pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safertrade"
version = "2.0.0"
description = "Advanced cryptocurrency trading intelligence platform"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SaferTrade Team", email = "contact@safertrade.org"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "web3>=6.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/safertrade/safertrade"
Documentation = "https://docs.safertrade.org"
Repository = "https://github.com/safertrade/safertrade"
Issues = "https://github.com/safertrade/safertrade/issues"

[project.scripts]
safertrade = "run_safertrade:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "engines*", "intelligence*", "shared*", "integrations*"]
exclude = ["tests*", "venv*", "logs*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
# Same as Black
line-length = 88
target-version = "py38"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "venv",
    "__pycache__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "archived",
    "logs",
    "htmlcov",
    "_paper",
    ".tools",
]

[tool.ruff.lint]
# Enable important rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "I",     # isort
    "N",     # pep8-naming
    "C4",    # flake8-comprehensions
    "PT",    # flake8-pytest-style
]
ignore = [
    "E501",  # Line too long (let formatter handle it)
    "E701",  # Multiple statements on one line (colon)
    "E722",  # Do not use bare except (keeping for now - will fix manually)
    "E731",  # Do not assign a lambda expression
    "E741",  # Ambiguous variable name
    "B006",  # Do not use mutable data structures for argument defaults
    "B008",  # Do not perform function calls in argument defaults
    "B904",  # Within an except clause, raise exceptions with raise ... from err
    "F401",  # Module imported but unused (may be used in tests/exports)
    "F841",  # Local variable assigned but never used
    "UP006", # Use `list` instead of `List` for type annotations (cosmetic)
    "UP007", # Use `X | Y` for type annotations (cosmetic)
    "UP035", # Import from `collections.abc` instead of `typing` (cosmetic)
    "W291",  # Trailing whitespace
    "W293",  # Blank line contains whitespace
    "SIM102", # Use a single if statement instead of nested if statements
    "SIM103", # Return the condition directly
    "SIM105", # Use contextlib.suppress instead of try-except-pass
    "SIM108", # Use ternary operator instead of if-else-block
    "SIM110", # Use all() instead of any()
    "SIM114", # Combine if branches using logical or
    "SIM115", # Use context handler for opening files
    "SIM117", # Use a single with statement for multiple contexts
]

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Tests can use assert statements and relative imports
"tests/**/*.py" = ["S101", "TID252"]
# Scripts can have print statements
"scripts/**/*.py" = ["T201"]

[tool.ruff.format]
# Like Black, use double quotes for strings
quote-style = "double"
# Like Black, indent with spaces, rather than tabs
indent-style = "space"
# Like Black, respect magic trailing commas
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending
line-ending = "auto"

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "archived", "logs", "_paper"]
skips = ["B101", "B601"]  # B101: assert_used, B601: paramiko_calls
